+++ packages/twenty-zapier/src/utils/handleQueryParams.ts
@@ -1,4 +1,4 @@
-import { InputData } from '../utils/inputData.type';
+import { InputData } from '../utils/data.types';
 
 const handleQueryParams = (inputData: InputData): string => {
   const formattedInputData: InputData = {};
